+++ api/entries.json
@@ -10,7 +10,7 @@
     {
       "entry_date": "2019-05-30",
       "entry_title": "Presenting our first group projects",
-      "entry_text": "Over the past few days we have been working on our group projects. We had to focus on the github flow and working on github with a team. Only HTML and CSS were allowed to be used. I am very proud of how our team worked together, and more than happy with how our website came out. Note to self, flexbox or floats, but not",
+      "entry_text": "Over the past few days we have been working on our group projects. We had to focus on the github flow and working on github with a team. Only HTML and CSS were allowed to be used. I am very proud of how our team worked together, and more than happy with how our website came out. Note to self, flexbox or floats, but not both.",
       "entry_mood": "Confident",
       "id": 2
     },
