+++ src/scripts/events.js
@@ -1,9 +1,23 @@
-// //--Radio Filter --//
+import { API } from "./data.js";
+import { RENDER, journalLog } from "./entriesDOM.js";
 
-// radioBtns = document.getElementsByName("moodFilter");
+const EVENTS = {
+  moodFilter: function() {
+    const radioBtns = document.getElementsByName("moodFilter");
 
-// radioBtns.ForEach(radioBtn => {
-//     radioBtns.addEventListener("click", () => {
-//         const mood = event.target.value
-//     })
-// });
+    radioBtns.forEach(radioBtn => {
+      radioBtn.addEventListener("click", event => {
+        const mood = event.target.value;
+        API.getJournalEntries().then(entries_obj => {
+          journalLog.innerHTML = "";
+          let filteredEntry = entries_obj.filter(
+            entry => entry.entry_mood === mood
+          );
+          RENDER.insertComponent(filteredEntry);
+        });
+      });
+    });
+  }
+};
+
+export { EVENTS };
