+++ src/styles/journal.css
@@ -5,39 +5,81 @@
 }
 
 body {
-  font-family: Helvetica, Arial, sans-serif
+  font-family: Helvetica, Arial, sans-serif;
+  background: slategray;
+}
+
+.main_container {
+  width: 1200px;
+  margin: auto;
+  background: white;
 }
 
 .header_h1 {
-  margin: 20px 0 20px 15px;
+  margin: 0px 0 20px 15px;
+  padding: 35px 35px 20px 35px;
 }
 
 /* Style Buttons */
-button {
+
+.submit-btn {
   height: 50px;
-  width: 120px;
+  width: 175px;
   border-radius: 10px;
   margin: 10px;
+  border-style: none;
+  background-color: slategray;
+  color: white;
+  font-weight: bold;
 }
 
-
+button {
+  height: 40px;
+  width: 75px;
+  border-radius: 10px;
+  margin: 10px;
+  border-style: none;
+  background-color: slategray;
+  color: white;
+  font-weight: bold;
+}
 
 /* Journal Entry Input Box */
 
-form {
-  margin-top: 20px;
-  margin-left: 20px;
+.input_container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.main_form {
+  font-weight: bold;
+  font-size: 1em;
+  border: 3px double slategray
 }
 
 fieldset {
   display: flex;
   flex-direction: column;
-  width: 600px;
+  width: 800px;
   padding: 10px;
+  border: none;
 }
 
 label {
   padding: 5px;
+}
+
+input {
+  border-radius: 3px;
+  width: 100%;
+  height: 30px;
+}
+
+textarea {
+  border-radius: 3px;
+  width: 100%;
+  height: 150px;
 }
 
 /* Journal Entry EDIT Input Box */
@@ -83,7 +125,11 @@ h1 {
   flex-wrap: wrap;
 }
 
-/* Modify Filter */
+.entry_text {
+  word-wrap: break-word;
+}
+
+/* Mood Filter */
 
 .moodContainer {
   display: flex;
@@ -91,12 +137,21 @@ h1 {
   justify-content: space-around;
 }
 
+.filter_form {
+  margin-top: 20px;
+  margin-left: 20px;
+}
+
+.filter_fieldset {
+  display: flex;
+  flex-direction: column;
+  width: 800px;
+  padding: 10px;
+  border: 1px solid black;
+}
+
 .moodItem {
   display: flex;
   align-items: center;
   width: 50px;
-}
-
-.entry_text {
-  word-wrap: break-word;
 }
